Make handling of Girouette return vals more general
Most of the time we get a [".class-name" {styling}] tuple, but we may also get a
CSSAtValue, or possibly any number of special Garden things. So instead of
special-casing these patterns, we walk whatever we get, replacing the classname
of the current token with `:&`. In a second pass we unpack the pattern `[:&
styling]`, this isn't necessary, but helps create more compact CSS.